:truck: Move FileMetadata types to file
